+++ packages/earn-controller/src/types.ts
@@ -1,0 +1,13 @@
+export type RefreshStakingEligibilityOptions = {
+  address?: string;
+};
+
+export type RefreshPooledStakesOptions = {
+  resetCache?: boolean;
+  address?: string;
+};
+
+export type RefreshPooledStakingDataOptions = {
+  resetCache?: boolean;
+  address?: string;
+};